Most object structure updates are done in a command chain; update object data on changes
svn path=/trunk/rkward/; revision=881

+++ rkward/dataeditor/celleditor.cpp
@@ -104,7 +104,7 @@ void CellEditor::keyPressEvent (QKeyEvent *e) {
 				return;
 			}
 		} else if (e->key () == Qt::Key_Right) {
-			if (cursorPosition () >= text ().length ()) {
+			if (cursorPosition () >= (int) text ().length ()) {
 				table->keyPressEvent (e);
 				return;
 			}
